+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## Unreleased
+
+- Keep the configured name case on append/add headers.
+  Fixes [#28](https://github.com/Kong/kong-plugin-request-transformer/issues/28)
+- Include the `type` function in template environment.
+  Because headers may contain array elements such as duplicated headers,
+  `type` is a useful function in these cases.
+- Fix the construction of the error message when a template throws a Lua error.
+  [#25](https://github.com/Kong/kong-plugin-request-transformer/issues/25)
+
 ## 0.37.4
 
 - chore(*) add copyright